renamed footer folder
