+++ src/components/footer/index.ts
