+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export * from "./container";
 export * from "./error-boundary";
-export * from "./footer2";
+export * from "./footer";
 export * from "./header";
 export * from "./social-icons";
